+++ R/qc_positive_control.R
@@ -8,8 +8,12 @@
 #'
 #' @return [[numeric]]
 qc_positive_control <- function(counts) {
-  measured <- log2(counts[["Count"]])
+  if (any(counts[["Count"]] %in% 0)) {
+    measured <- log2(counts[["Count"]] + 1)
+  } else {
+    measured <- log2(counts[["Count"]])
+  }
   known <- log2(as.numeric(sub("^[^(]*\\((.*)\\)$", "\\1", counts[["Name"]]))) # plexset value: "32"
-  correlation <- stats::cor.test(known, measured)$estimate
+  correlation <- summary(stats::lm(measured ~ known))$r.squared
   unname(round(correlation, 5))
 }